+++ src/Fantomas.Core/SourceParser.fs
@@ -86,7 +86,6 @@ let (|Open|_|) =
 
 let (|OpenType|_|) =
     function
-    // TODO: are there other SynType causes that need to be handled here?
     | SynModuleDecl.Open (SynOpenDeclTarget.Type (SynType.LongIdent synLid, _m), _) -> Some synLid
     | _ -> None
 
@@ -227,9 +226,9 @@ let (|UnionCaseType|) =
     | SynUnionCaseKind.Fields fs -> fs
     | SynUnionCaseKind.FullType _ -> failwith "UnionCaseFullType should be used internally only."
 
-let (|Field|) (SynField (ats, isStatic, ido, t, isMutable, px, ao, range)) =
+let (|Field|) (SynField (ats, _isStatic, ido, t, isMutable, px, ao, range, { LeadingKeyword = lk })) =
     let innerRange = ido |> Option.map (fun i -> Range.unionRanges i.idRange t.Range)
-    (ats, px, ao, isStatic, isMutable, t, ido, innerRange, range)
+    (ats, px, lk, ao, isMutable, t, ido, innerRange, range)
 
 let (|EnumCase|) (SynEnumCase (ats, ident, c, cr, px, r, trivia)) =
     let fullRange =
@@ -286,13 +285,27 @@ let (|MDMember|_|) =
 
 let (|MDLetBindings|_|) =
     function
-    | SynMemberDefn.LetBindings (es, isStatic, isRec, _) -> Some(isStatic, isRec, es)
+    | SynMemberDefn.LetBindings (bindings = es) -> Some es
     | _ -> None
+
+let mkWithGetSet t withKeyword memberKind =
+    let isFunctionProperty =
+        match t with
+        | Some (SynType.Fun _) -> true
+        | _ -> false
+
+    withKeyword
+    |> Option.map (fun _ ->
+        match memberKind with
+        | SynMemberKind.PropertyGet -> if not isFunctionProperty then "" else " with get"
+        | SynMemberKind.PropertySet -> " with set"
+        | SynMemberKind.PropertyGetSet -> " with get, set"
+        | _ -> "")
 
 let (|MDAbstractSlot|_|) =
     function
-    | SynMemberDefn.AbstractSlot (SynValSig (ats, ident, tds, t, _, _, _, px, ao, _, _, _), mf, _) ->
-        Some(ats, px, ao, ident, t, tds, mf)
+    | SynMemberDefn.AbstractSlot (SynValSig (ats, ident, tds, t, _, _, _, px, ao, _, _, trivia), mf, _) ->
+        Some(ats, px, ao, trivia.LeadingKeyword, ident, t, tds, mkWithGetSet (Some t) trivia.WithKeyword mf.MemberKind)
     | _ -> None
 
 let (|MDInterface|_|) =
@@ -302,21 +315,19 @@ let (|MDInterface|_|) =
 
 let (|MDAutoProperty|_|) =
     function
-    | SynMemberDefn.AutoProperty (ats,
-                                  isStatic,
-                                  ident,
-                                  typeOpt,
-                                  mk,
-                                  memberFlags,
-                                  _memberFlagsForSet,
-                                  px,
-                                  ao,
-                                  equalsRange,
-                                  e,
-                                  withKeyword,
-                                  _,
-                                  _) ->
-        Some(ats, px, ao, mk, equalsRange, e, withKeyword, ident, isStatic, typeOpt, memberFlags)
+    | SynMemberDefn.AutoProperty (ats, isStatic, ident, typeOpt, mk, _, _, px, ao, e, _, trivia) ->
+        Some(
+            ats,
+            px,
+            ao,
+            trivia.LeadingKeyword,
+            trivia.EqualsRange,
+            e,
+            mkWithGetSet typeOpt trivia.WithKeyword mk,
+            ident,
+            isStatic,
+            typeOpt
+        )
     | _ -> None
 
 let (|MDPropertyGetSet|_|) =
@@ -325,8 +336,8 @@ let (|MDPropertyGetSet|_|) =
                                                     isInline = isInline
                                                     attributes = ats
                                                     xmlDoc = px
-                                                    valData = SynValData(memberFlags = Some mf)
-                                                    headPat = SynPat.LongIdent (longDotId = memberName)) as getBinding),
+                                                    headPat = SynPat.LongIdent (longDotId = memberName)
+                                                    trivia = { LeadingKeyword = lk }) as getBinding),
                                   Some setBinding,
                                   _,
                                   { GetKeyword = Some getKeyword
@@ -341,14 +352,14 @@ let (|MDPropertyGetSet|_|) =
                 (SynMemberDefn_GetSetMember_Set, setKeyword, setBinding),
                 Some(SynMemberDefn_GetSetMember_Get, getKeyword, getBinding)
 
-        Some(px, ats, mf, isInline, ao, memberName, withKeyword, firstBinding, andKeyword, secondBinding)
+        Some(px, ats, lk, isInline, ao, memberName, withKeyword, firstBinding, andKeyword, secondBinding)
     | SynMemberDefn.GetSetMember (None,
                                   Some (SynBinding (accessibility = ao
                                                     isInline = isInline
                                                     attributes = ats
                                                     xmlDoc = px
-                                                    valData = SynValData(memberFlags = Some mf)
-                                                    headPat = SynPat.LongIdent (longDotId = memberName)) as binding),
+                                                    headPat = SynPat.LongIdent (longDotId = memberName)
+                                                    trivia = { LeadingKeyword = lk }) as binding),
                                   _,
                                   { WithKeyword = withKeyword
                                     GetKeyword = getKeyword
@@ -357,8 +368,8 @@ let (|MDPropertyGetSet|_|) =
                                                     isInline = isInline
                                                     attributes = ats
                                                     xmlDoc = px
-                                                    valData = SynValData(memberFlags = Some mf)
-                                                    headPat = SynPat.LongIdent (longDotId = memberName)) as binding),
+                                                    headPat = SynPat.LongIdent (longDotId = memberName)
+                                                    trivia = { LeadingKeyword = lk }) as binding),
                                   None,
                                   _,
                                   { WithKeyword = withKeyword
@@ -369,7 +380,7 @@ let (|MDPropertyGetSet|_|) =
             Some(
                 px,
                 ats,
-                mf,
+                lk,
                 isInline,
                 ao,
                 memberName,
@@ -382,7 +393,7 @@ let (|MDPropertyGetSet|_|) =
             Some(
                 px,
                 ats,
-                mf,
+                lk,
                 isInline,
                 ao,
                 memberName,
@@ -446,24 +457,36 @@ let parseExpressionInSynBinding returnInfo expr =
         e
     | _ -> expr
 
-let (|DoBinding|LetBinding|MemberBinding|ExplicitCtor|ExternBinding|) b =
-    match b with
+let (|Binding|) (SynBinding (ao, _, isInline, isMutable, attrs, px, _, pat, returnInfo, expr, _, _, trivia) as b) =
+    let e = parseExpressionInSynBinding returnInfo expr
+    let rt = Option.map (fun (SynBindingReturnInfo (typeName = t)) -> t) returnInfo
+    attrs, px, trivia.LeadingKeyword, ao, isInline, isMutable, pat, rt, trivia.EqualsRange, e, b.FullRange
+
+let (|ExplicitCtor|_|) =
+    function
     | SynBinding (ao, _, _, _, ats, px, SynValData (Some MFConstructor, _, ido), pat, _, expr, _, _, trivia) ->
-        ExplicitCtor(ats, px, ao, pat, trivia.EqualsRange, expr, ido)
-    | SynBinding (ao, _, isInline, _, ats, px, SynValData (Some mf, _, _), pat, returnInfo, expr, _, _, trivia) ->
-        let e = parseExpressionInSynBinding returnInfo expr
-        let rt = Option.map (fun (SynBindingReturnInfo (typeName = t)) -> t) returnInfo
-        MemberBinding(ats, px, ao, isInline, mf, pat, rt, trivia.EqualsRange, e)
-    | SynBinding (_, SynBindingKind.Do, _, _, ats, px, _, _, _, expr, _, _, _trivia) -> DoBinding(ats, px, expr)
-    | SynBinding (ao, _, _, _, attrs, px, _, pat, returnInfo, _, _, _, { ExternKeyword = Some externRange }) ->
+        Some(ats, px, ao, pat, trivia.EqualsRange, expr, ido)
+    | _ -> None
+
+let (|DoBinding|_|) =
+    function
+    | SynBinding (_, SynBindingKind.Do, _, _, ats, px, _, _, _, expr, _, _, trivia) ->
+        Some(ats, px, trivia.LeadingKeyword, expr)
+    | _ -> None
+
+let (|ExternBinding|_|) =
+    function
+    | SynBinding (accessibility = ao
+                  attributes = attrs
+                  xmlDoc = px
+                  headPat = pat
+                  returnInfo = returnInfo
+                  trivia = { LeadingKeyword = SynLeadingKeyword.Extern _ as lk }) ->
         let rt =
             Option.map (fun (SynBindingReturnInfo (typeName = t; attributes = a)) -> a, t) returnInfo
 
-        ExternBinding(externRange, ao, attrs, px, pat, rt)
-    | SynBinding (ao, _, isInline, isMutable, attrs, px, _, pat, returnInfo, expr, _, _, trivia) ->
-        let e = parseExpressionInSynBinding returnInfo expr
-        let rt = Option.map (fun (SynBindingReturnInfo (typeName = t)) -> t) returnInfo
-        LetBinding(attrs, px, trivia.LetKeyword, ao, isInline, isMutable, pat, rt, trivia.EqualsRange, e)
+        Some(lk, ao, attrs, px, pat, rt)
+    | _ -> None
 
 // Expressions (55 cases, lacking to handle 11 cases)
 
@@ -864,46 +887,22 @@ let (|JoinIn|_|) =
 /// Recursive and use properties have to be determined at this point
 let rec (|LetOrUses|_|) =
     function
-    | SynExpr.LetOrUse (isRec, isUse, xs, LetOrUses (ys, e), _, trivia) ->
-        let prefix =
-            if isUse then "use "
-            elif isRec then "let rec "
-            else "let "
-
+    | SynExpr.LetOrUse (_, _, xs, LetOrUses (ys, e), _, trivia) ->
         let xs' =
             let lastIndex = xs.Length - 1
-
-            List.mapi
-                (fun i x ->
-                    if i = 0 then
-                        (prefix, x, (if i = lastIndex then trivia.InKeyword else None))
-                    else
-                        ("and ", x, (if i = lastIndex then trivia.InKeyword else None)))
-                xs
+            List.mapi (fun i x -> if i = lastIndex then x, trivia.InKeyword else x, None) xs
 
         Some(xs' @ ys, e)
-    | SynExpr.LetOrUse (isRec, isUse, xs, e, _, trivia) ->
-        let prefix =
-            if isUse then "use "
-            elif isRec then "let rec "
-            else "let "
-
+    | SynExpr.LetOrUse (_, _, xs, e, _, trivia) ->
         let xs' =
             let lastIndex = xs.Length - 1
-
-            List.mapi
-                (fun i x ->
-                    if i = 0 then
-                        (prefix, x, (if i = lastIndex then trivia.InKeyword else None))
-                    else
-                        ("and ", x, (if i = lastIndex then trivia.InKeyword else None)))
-                xs
+            List.mapi (fun i x -> if i = lastIndex then x, trivia.InKeyword else x, None) xs
 
         Some(xs', e)
     | _ -> None
 
 type ComputationExpressionStatement =
-    | LetOrUseStatement of prefix: string * binding: SynBinding * inKeyword: range option
+    | LetOrUseStatement of binding: SynBinding * inKeyword: range option
     | LetOrUseBangStatement of isUse: bool * SynPat * equalsRange: range option * SynExpr * range: range
     | AndBangStatement of SynPat * equalsRange: range * SynExpr * range: range
     | OtherStatement of SynExpr
@@ -1276,7 +1275,7 @@ let (|SimplePats|SPSTyped|) =
 
 let (|RecordField|) =
     function
-    | SynField (ats, _, ido, _, _, px, ao, _) -> (ats, px, ao, ido)
+    | SynField (ats, _, ido, _, _, px, ao, _, { LeadingKeyword = lk }) -> (ats, px, lk, ao, ido)
 
 let (|Clause|) (SynMatchClause (p, eo, e, range, _, trivia)) =
     let fullRange =
@@ -1593,7 +1592,9 @@ let (|TyparSingle|TyparDefaultsToType|TyparSubtypeOfType|TyparSupportsMember|Typ
 
 let (|MSMember|MSInterface|MSInherit|MSValField|MSNestedType|) =
     function
-    | SynMemberSig.Member (vs, mf, _) -> MSMember(vs, mf)
+    | SynMemberSig.Member (vs, mf, _) ->
+        let (SynValSig (synType = t; trivia = trivia)) = vs
+        MSMember(vs, mkWithGetSet (Some t) trivia.WithKeyword mf.MemberKind)
     | SynMemberSig.Interface (t, _) -> MSInterface t
     | SynMemberSig.Inherit (t, _) -> MSInherit t
     | SynMemberSig.ValField (f, _) -> MSValField f
@@ -1602,7 +1603,12 @@ let (|MSMember|MSInterface|MSInherit|MSValField|MSNestedType|) =
 let (|Val|)
     (SynValSig (ats, synIdent, SynValTyparDecls (typars, _), t, vi, isInline, isMutable, px, ao, eo, range, trivia))
     =
-    (ats, px, trivia.ValKeyword, ao, synIdent, t, vi, isInline, isMutable, typars, eo, range)
+    let synIdent =
+        match trivia.LeadingKeyword with
+        | SynLeadingKeyword.New _ -> None
+        | _ -> Some synIdent
+
+    (ats, px, trivia.LeadingKeyword, ao, synIdent, t, vi, isInline, isMutable, typars, eo, range)
 
 // Misc