+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 * Formatting of anonymous record types respects fsharp_multiline_block_brackets_on_same_column. [#1167](https://github.com/fsprojects/fantomas/issues/1167)
-* Update FCS to 'Add SynType.Or', commit b78ad5cd6feea006986869506f33aab84c37398a
+* Update FCS to 'Leading keyword synbinding', commit 05201993cfaac663c72b6431d81f543e3d001dfe
 
 ## [5.1.0-alpha-005] - 2022-10-07